fix warnings

+++ include/lexer.h
@@ -27,7 +27,7 @@ public:
 		return buffer.substr(range.begin, range.end - range.begin);
 	}
 	[[nodiscard]] size_t position() const { return ptr - buffer.data(); }
-	[[nodiscard]] char peek() const { return (ptr - buffer.data() >= buffer.size())? (char) EOF: *ptr; }
+	[[nodiscard]] char peek() const { return (static_cast<size_t>(ptr - buffer.data()) >= buffer.size())? (char) EOF: *ptr; }
 	void rewind_to(size_t pos) { ptr = buffer.data() + pos; }
 	char take() {
 		char r = peek();
